added XmlValidatorFilter and XmlValidator actor, the latter one being a filter based router that can be used to collect statistics about messages that do not conform to a schema. Also added respective regression tests
added filtering to KafkaImportActor, to keep it similar to JmsImporter

added basic support for WaveFront (obj) meshes. This requires a new WorldWindJava-pcm that includes the respective
forum contributions from robotfire. Added a FlightModel as the rendering abstraction

started FlightLayer3D that maintains a proximity list for the current eye position, and switches to a model in case
flights are close enough

Added a MinClipOrbitView to minimize far clip distance in case the eye position is close enough to the globe that we
don't see a horizon, in which case we can reduce the far clip significantly and hence also reduce the near clip so that
we don't clip out proximity list models. This unveiled a problem with flicker at far distance if the near clip is too low
(e.g. with WWJs default min clip).