fix for vanishing tracks. Repeat 100x "Double.NaN != Double.NaN". The underlying problem is that we get asdex positionReports that don't have both lat & lon, and the isDefined guard was failing
fixed ChannelTopic problem that was caused by having a default handleMessage in FilteringPublisher, which
was matching all BusEvents, and we also used BusEvent for CT messages. We do want to support the use of
channels for system messages, hence Bus is now parameterized with a more generic ChannelMessage, of which
we have two sub-types: user-level BusEvent and system-level BusSysEvent. The latter one is used by
ChannelTopicSubscriber/Provider.

NOTE - in the wake of this change BusEvent got moved to Messages, which requires all its users to be
change imports. This is why we have to bump up the minor version to 1.4.0 since all external clients
would have to be updated accordingly

Replaced websocket use in http TestRoutes - there is no value in using the akka-streams overhead
if all that is transmitted is a simple GET, and the client needs to have the full response before
refreshing the page. Example now uses simple client side setInterval() and XMLHttpRequest, and server
(route) side uses simple path/complete directives