FilterActor was no subscriber
re-instated logback-console.xml so that we can use a LogConsole. This is
just a workaround though since we need the RaceLogger for pre-actor
logging, hence it should be able to write to a socket

fixed ProcessLancherSpec to avoid hardcoded scala version paths

extended swim-sfdps-stats.conf to use a XmlValidator and also count
rejected messages. It also contains a FilterActor to avoid large
number of FDPSMsg messages that are not in the FIXM schema