added more FPosStatsCollector checks, to detect stale (old) incoming flights, flights that are not updated within a configured interval (ala FlightDropper, identical (duplicated) flight positions, and - most notably - ambiguous flight positions. The latter ones happen if several stations report the same flight with <1sec sample time differences, since SFDPS messages seem to truncate fractional seconds.
The reason why not to use the FlightDropperActor for dropped flights is
that we also want to get stale incomings (separate count), and since we
already have an active flight collection and a timer the
FlightDropperActor would be too redundant.

Added a protocol to request a SimClock time reset. The motivation is
that currently sfdps has only historical data, but we still want to
check for consistency within the sample set, hence we have to set the
sim time from the first received position. Resetting the sim clock is a
serious business though, since it might affect a number of other actors.
For that reason, it has to be (non-overridable) configured by setting
'can-reset-clock' in the respective actor to true. Requests have to be
submitted to the RAS through
ContinuousTimeRaceActor.resetSimClockRequest(d), which then sends a
respective RaceResetClock to the master, which send ResetSimClock
messages to the actors (resulting in overridable onSimClockReset()
invocations). Actors respond with SimClockReset, master responds with
RaceClockReset.

WATCH OUT - all this is synchronous, to avoid actors running
undetected with different time bases.

Removed the separate logging from XmlValidator with an optional failure
prefix/postfix that get prepended before writing the message to the
write-to-fail channel. That way, we can record failed messages plus
failure reason with a simple TextArchiveWriter. At some point we should
still switch to AnnotatedItems though, to avoid extra allocation.

Changed ConsoleMain to not attempt to instantiate a RAS if there was a
vault set but the user entered the wrong or no password. This requires
ConfigVault.initialize to return a boolean