changed onXX transition callbacks to return boolean, indicating if the transition was successfully completed or failed. This also means that the standard pattern is to either use the super method return value at the end, or use it in a conditional execution of the body in case the class depends on super state. This is required to fully implement a success/failure/exception behavior
refactored sim clock update, which has to happen from inside of master
since the master has to revert changes if they get rejected by one of
the actors (a consistent clock is a primary asset). The requesting actor
now directly sends the sync request to the master, and is exempt from
SyncWithRaceClock notification processing since it would run into a
timeout if the master does such notifications synchronously (this is
from an ask pattern inside the requester). We now explicitly call the
onSyncWithRaceClock of the requester if the master succeeds notifying
all other actors.

This still needs to be unified with inter-RAS clock synchronization!

Refactored HtmlStats/HttpStatsReporter to obtain not only html but also
potential resources such as images from the rendered stats objects. We
use a extractUnmatchedPath directive in the reporter route to look up
HttpEntities in the map collected from the stats topics. Also
added configurable formatters in addition to direct HtmlStats
implementors since it is quite likely that we have more than one
format for a stats topic. Added a jfreechart based image example for
FPosStats (histogram). This should however use SVG instead of a
somewhat blurred PNG rendering