refactored high update frequency flight path rendering, introducing the beginnings of a TimeSeries trait hierarchy
fixed FlightLayerInfoPanel bug that stopped the timer which is used to
update the match list. DynamicLayerInfoPanel was stopping the timer if
the whole panel is hidden. A better approach is to not update fields if
the panel is not showing

fixed RaceActorSpec, TRAS shutdown was runing into timeouts. The
TestMaster should override the receive of the Master

implemented a simple TypedRouter, which can be used to map (BusEvent)
message types to channels. This is going to be used as a sequencer to
investigate ADS-B backumps. Also added a regression test