changed FlightPos from case class to regular class since we want to be able to derive specialized types
refactored the result management of XmlParser. We need both - a _result var for parsers that set it during their onStart/EndElement execution, and a result method for parsers that collect the result when all of the relevant text has been parsed. The default implementation just returns _result

fixed FIXM2FlightObject parser, which fell victim to the XmlParser 'result' change