made RAS timeouts (both system and per-actor) explicit and configurable (create-timeout,init-timeout,start-timeout). Replay of large archives that have to skip a lot of records can otherwise run into timeouts
added TaisDWArchiveReader and sample config to replay

added a convenience task to compile Avro schemas (*.avsc), which does
not follow the more expensive approach of plugins such as sbt-avrohugger
(compiling scala into $sourceManaged) since these are based on Avro
versions that are incompatible with latest.release. Since we value
up-to-date production libs higher, we rely on avro-tools to be installed
externally and add the generated files to the repo (src/main/java) so
that not all users need the tools. This follows the same approach as
DDS IDL generated files.

added simple Apache Avro support to gov.nasa.race.archive (we only use
Avro for archives since it stores the schema in the respective archive,
which makes it un-suitable as a wire format for high frequent messages.

added a roundtrip reg test for Avro