refactored AsdexMsg translation. The parsing is now in the root, which also doubles as raw AsdexTrack generator (i.e. no accumulation of delta info). AsdexMsg2FullAsdexTracks is the accumulating translator that caches the track info, which also comes in handy in case of zooming in/out of airports without loosing all the accumulated data
added undefined values plus Option-like APIs to uom types (Angle, Speed, Length). This is helpful for unifying tracks since positions etc. are not optional in all but raw AsdexTrack updates. It also is more efficient for local var caches in parsers, to avoid heap pressure

finally made Scala version a normal CommonRaceSettings val, so that it can be referenced in library dependencies that are Scala version specific (e.g. scala-reflect)