big ArchiveReader refactor, which now is again more scalatic but still avoids gazillions of short living objects
DWArchiveReader now is heavily leaning on the new SearchStrean, which is less code and more efficient since it does less buffer copying. This also solves the
switching-airport bug when replaying asdex archives (which was caused by not properly separating messages in the old implementation)

added test applications for both SearchStream and DWArchiveReader. Those are just applications because all the bugs only showed in
input files > 1G, which we can't add to the repo. Maybe we should support machine configured reg tests