refactored and fixed TFMTrackInfoParser, which now also does (some) waypoint parsing (not complete yet)
changed XmlParser default behavior for whileNextElement to ignore
match exceptions since it is too easy for clients to forget to check
parse errors, which then results in silently ignored data. If a client
wants to cause parse errors it has to do so explicitly. The normal use
case is to just define cases that are handled.

first version of a new TextArchiveExtractor tool to extract single
messages from large archives. Comes in handy for debugging. Added an
interactive TFMTrackInfoParserTest to make use of it

renamed .../avro/TrackInfo to TrackInfoRecord. Avoids import renames
since it is mostly used in a context that also refers to the regular
TrackInfo

fixed the RaceViewer view construction to happen from the UI thread.
This happens now sync from the InitializeRaceActor. Fixed the
"uncommitted CA transactions" bug under newer OS X versions