moved BufferRecord back to having a invariant buffer since it doesn't make sense trying to keep/update a single Record instance in mapped CLBuffers. Host can only access when mapped, device only when unmapped. Added executeMapped() methods that encapsulate the mapping scope and provide the required access object (ByteBuffer or BufferRecord) to the client
added more BufferRecord field types

added a sliced ByteBuffer that only contains the record data, which we need if we want to do clEnqueueWriteBuffer calls for single (changed) records. Not clear at this point though where the break even is in terms of number of clEnqueueWriteBuffer calls vs. size of transferred memory.

Made CLCommandQueue.enqueueWrite non-blocking