on OS X, clCreateContext fails with a misleading CL_INVALID_VALUE if the platform is not explicitly specified in ctxProps
added homogeneity checks for platform and byte order to CLContext. While
the first one is a no-brainer, the byte order is a constraint we
introduce. However, we have to make sure we use the right byte order for
direct buffers and it seems very error prone if this has to be
dynamically set based on which queue/device we use for the same buffer