fix for the AsdexTrack problem with detecting/remembering aircraft tracks - Asdex airports don't set the aircraft flag consistently, and since it should be a sticky attribute we store it in the AsdexEntry
fixed the AsdexLayer problem with setting symbols - if initializeLayer
is overridden it has to call super (or replace respective actions)

beginnings of 3D flight paths (should be a display attribute set from
the object pane). This still lacks support for perspective of the
aircraft symbols (we can't use PointPlacemark if there is a view pitch).

fix for the actor-based termination requests (RaceViewer). This is sent
to the master, which in turn used ras.terminationRequest to
check/initiate proper termination sequence. Default should be disabled
self-terminatio though