made the FHT{1,2,3}Interpolants independent of the actual observation storage, which is now abstracted by function parameters. This can be used to interpolate observations which are kept in collections other than simple 0-based arrays (e.g. in circular buffers)
added a generic FHTnInterpolant, which allows us to interpolate any
number of dimensions. This makes use of that we only need to compute
one set of weights for FH interpolants, which in turn can be a huge
benefit if we have to handle (extensible) state vectors

added FHT1 class as convenience type for simple array based FHTinterpolant

updated FHT1InterpolantSpec to use FHT1 instead of the generic
FHT1Interpolant

fixed the selTracon initialization in TATrackLayer, which prevented it
from being configured via 'request-topics'

added proper sublabel updates in TrackEntry/TrackSymbol

fixed the local var masking of TATrackEntry 'obj' - beware of the
common Scala pitfall of var masking by means of (non-updated) local ctor
args. ALWAYS USE A '_<var>' PREFIX to avoid this.