first stab at CompressedTrajectory, using DeltaDateArray and LatLonArray. Still needs some refactoring to compile, snd still has too much redundancy in trace implementations
OffsetTraj still needs to use uom arrays

unfortunately those have a substantial amount of duplicated code, which can not be avoided by means of generic base because of performance penalties for element type conversion functions.

note that Time now wraps Int, limiting durations to <=24d. Unfortunately we can't have a generic value class with a context bound (T : Integral) since that introduces a implicit parameter, which is not allowed for value class ctors (compile error is not too intuitive)