first pass over complete set of trajectory implementations, including accurate trajectories
rework of uom arrays with defined interface. Unfortunately we need to
have separate types for the underlying element types
(Long,Float,Double), of which there is only UOMDoubleArray/Buffer so
far. The reason is that Scala does not allow context bounds on trait
type parameters, which we need in order to do arithmetic operations on
elements

added delta arrays for Length, Angle