refactored LayerInfoPanels, adding a explicit SharedLayerInfoPanel for the case of (WW) layers that don't have their own panels. The panels for subscribing layers now get their layer from a invariant ctor arg (which eliminates inconsistent setLayer() calls.
Also added a StaticLayerInfoPanel, which assumes a fixed number of
LayerObject entries that is known at panel construction

RaceLayers are supposed to hold LayerObjects, and report the number with
a size() method.

turned LayerObjectAttribute and LayerObjectAction into plain Scala enums

added attribute change to InteractiveRaceLayer, to stay consistent (all
rendering relevant changes should go through the layer since it is
responsible for updating object/info panels and sending out change
messages. It is still suboptimal that we need to arm the
raceViewer.trackUserAction before we know if there is something to send
- this should be in the same place. However, this would require to
distinguish between interactive and automatic changes in respective
layers)

Updated path/info/mark attribute changes accordingly (still more to be
done)