re-instated close check in TrackDiffActor for diff trajectories that don't get updated, although it's debatable if this should not just be left for a FlightDropperActor (which might be required anyways if we display the respective channel in a TrackLayer)
deferred cleanup of entries in TrackDiffActor to avoid re-entering a new
record once the track has left the zone into the capture area

various checks for empty/single point trajectories, which otherwise
cause lots of exceptions for delta computations

make simClock timeScale available in RaceActor so that deferred actions
can be based on simTime instead of wall time

fixed bug in FIXM2FlightObject which was not properly exiting the flight
element matcher and hence not re-initializing parsed variables. This
could have caused wrong values for optional elements/attributes

exteded TrackPairEventLayer to support labels and colors for trackpos
markers and paths. Both are optional