generified MurMurHash64.hash() over Byte,Char but not clear if extra runtime cost is acceptable. Still needs to properly handle non-ASCII Array[Char] and Strings
added a UTF8 support object which can be used to compute utf-8 length
of Array[Char] slices and Strings without allocating/encoding. This also
includes a value class based Encoder which can be used to iterate over
the resulting bytes without requiring allocation. This will be the basis
for computing utf-8 based hashes for Array[Char] and String