opt:l:inline is to be used cautiously since it leads to non-reproducible results when compiling incrementally, and it should only be used within RACE packages or it can result in Java runtime errors
ad hoc fixes for a number of @inline specs that were not final. There
are a gazillion more

More attempts to make streaming akka-http gracefully terminate in a
non-streaming HttpImportActor. Still might cause warnings.

Also added a 'coalesce' option to HttpImportActor in case we don't need
new requests if there still is a pending one.