turned RaceTick into RaceSystemMessage and gave it a onRaceTick so that we don't need to override handleMessage/handleSystemMessage just for the sake of processing periodic RaceTicks
reworked PeriodicRaceActor tickInterval/Delay initialization, which is
now performed during onRaceActorInitialize. This means we can't schedule
during construction but that is safer anyways. Also added the
automatic startSchedule to PeriodicRaceActor so that derived types don't
have to override onStartRaceActor just for this purpose

first version of SBSImportActor with integrated drop support via
PeriodicRaceActor. This can be still improved by moving the cleanup
check completely into the data acquisition thread and using soTimeout
reads to guarantee we would also cleanup in the unlikely case that no
input was read within the check interval