bumped SBT up to 1.3.5
changed/extended StringDataBuffer interface/implementation to support a
more StringBuilder like API (including adding single chars)

added globally unique TATrack ids, which are interned concatenations of
Tracon id and trackNum. Since the cs is still initialized from trackNum
if there is no acid this will not change the display

with that, the TAISImportActor is functional and can be used as a
template for other ChannelTopicProvider importers (ASDEX).