replace .geo.LatLon with simple fix point representation, which is easier to compute and has higher accuracy than the old grid+fixpoint scheme. This has about 1cm accuracy for lat and lon, which is enough for our purposes
replaced gov.nasa.race.common.inlined with explicit inline class list.
This should be used sparingly anyways, and we should be able to switch
from inline to non-inline without having to modify clients. The downside
is that it is not obvious anymore that clients use inlines, but at least
we stay consistent with respect to package names (which should not
reflect implementation)

updated Akka and a number of other external dependencies