bumped worldwind-pcm to r202, which is the fat jar including current snapshots of jogl and gluegen 2.4.0 from (https://jogamp.org/deployment/v2.4.0-rc-20200307)
once these make it to the central repository we remove them from
worldwind-pcm again. Publishing fat jars is in general a bad idea but
since the only client in a RACE app should be race-ww this is better
than either adding the jogl/gluegen jars as unmanaged dependencies or
by restricting the Java versions that can run respective RACE apps with
the last published artifacts (2.3.2 as of now). On macOS 10.14.6 that
was <= 12.0.1. All newer ones caused hard errors

This change allows to run RACE on Java up to 14.0. Unfortunately the
scalac optimizer does not work with the newer JDKs and method
optimization had to be disabled for now