fixed broken pause/resume, which now works from all toplevel RAs. There still is some redundancy with non-actor pause/resume and some inconsistency with termination that is legacy from pre-ParentActor master
generalized websocket RouteInfo traits

added missing protocol websocket support

generalized the (usually evil) case of setting data in non-actor objects from RAs. While this might require synchronization in the non-actor it allows to use the generic RaceDataConsumerActor in all consumers that directly accept the bus message payloads. Note however that we can't match on generic types for such payloads in the consumer because of type erasure. If the consumer needs data that is computed from generic types it might still be better to generate this in the associated actor, which also solves synchronization (see HttpStatsReporter)