add custom akka-http ServerSettings support. So far we only use this for a server option 'ws-keep-alive' that is an alias for akka.http.server.websocket.periodic-keep-alive-max-idle
use a partial function to implement websocket protocols which only handles
the messages it cares about, and consume the others in a Nil returning wrapper
in ProtocolWSRaceRoute. This adds some safety in case clients forget to
consume ignored messages (not clear yet how much though).

This leaves websocket support with transparent per-route keep alive (aka
Ping frames). While it can be simulated with adding a keepAlive stage to the flow
this is not transparent, i.e. the client has to know about such messages.