first stab at SiteRoute, including AuthorizedSiteRoute, which is a RaceRouteInfo for mixed static/dynamic content
split up RaceRouteInfo which was getting too big

fixed login/logout behavior of AuthorizedRaceRoute, which now uses the same URI (requestPrefix/{login,logout}) for both GET (explicit auth request) and POST (the actual user auth with credentials). This had to be a path so that client browsers send the  session token cookie (if there is one). Note that matching paths in akka-http has the pitfall of not working with single string literals and require PathMatchers

still needs clean up for route composition