added configurable app menu to ConsoleMain. This can be used as a more convenient way to send (String) commands to race actors
changed RaceActorSystem.send() to use actor name instead of selection path. Means we can only send to named RaceActors but that is probably what we should restrict ourselves to anyways

added a sim-mode config to UpstreamConnectorActor, which enables two command messages: "cut" to simulate connection loss and "restore" to revert this. Useful for demo purposes to show re-synchronization