the big user authentication overhaul part 1
this replaces the Authenticator with AuthMethod, which is refactored so
that it allows to configure the authentication method without changes to
the respective route that uses it.

there are two major modes for AuthMethods: (1) document request (running the
auth protocol through GET/response) and (2) websocket embedded. The
latter one is used to protect certain websocket operations (e.g. data
modification) and runs embedded in the websocket client message handler
of the route. Both modes share significant code, such as the server side
protocol.

single page websocket applications have to provide a "auth" layer
(iframe) to enable the websocket user auth support

still missing is a mode that uses the same authMethod for both document
and websocket access protection (i.e. remembers document user auth)