first stab at CesiumTrackRoute. Still missing support to proxy map tile server
added race-client-ui to have a common place for all the generic browser
UI code. Only JS/css so far but will eventually include a scala-tags
layer for generating html. The philosopy is to have manually
editable/testable HTML content that gets expanded/rewritten by means of
JS on the client, in order to support the type of dense UI we got from
Swing in the WorldWind viewer. This is still in its very beginnings