added a first version of a simple CachedProxyRoute. Still needs to be more robust WRT exceptions and currently needs to run on Java-17 to support user file attributes on macOS/HFS+
first stab at a proxying CesiumTrackRoute. The various configuration
levels for the client side still need some reshuffling, and so does the
server-side client asset management (which currently comes from class
resources)

added a tracks URL to the share-node_1.conf. Note this requires a vault
for the Cesium app key