temp fix for the flickering dynamic polyline/wall update, which appears to be a bug in Cesium's Polyline/WallGraphics implementation. Using a CallbackProperty to add new trajectory points does not display the wall, and apparently concatenates existing and new points in a polyline with the current position at the wrong end of the new path. Using just a normal ConstantProperty flickers on update as long as there is only one of the respective primitives. Workaround is to automatically add a minimal start segment of the trajectory with the same presentation parameters for each new entity
added a onselect callback for the list widget which does fire each time
the selection changes no matter if by user interaction or
programmatically. This simplifies using lists to set widgets controlling
aspects of the selected item. The old onclick is still there but should
be only used if the action is bound to interactive use of the list
widget