fixed ProxyRaceRoute caching, which was failing to store WMS tiles because of macOS filename length constraints. Now uses an md5 checksum for long filenames
added chunked transfer of large / unknown size files to CesiumGeoLayer (might have to be refactored since it is more general). While we shouldn't serve GB files it is quite common to have compressed files that we have to send de-compressed, which is the same problem. We have to be able to serve from a (finite) stream, not just strict content