next round of odin_actor refinement.
Firstly, we went back to tokio_{kanal,channel,..} - the channel trait abstraction turned out to be less readable (tokio_rt.rs) and also required runtime overhead since trait aliases don't support async traits, which is precisely what we need for the concrete Actor trait. In general, the additional level of abstraction seemed to introduce overly conservative type bounds, although that probably could have been solved with the generic Actor struct. Since it both increased runtime cost and decreased readability and in addition did only save 20% code over tokio_channel + tokio_kanal it was just not worth it.

The other big change is to move from an Actor trait to a generic Actor struct. This became necessary because of the Ask pattern support. The Ask message struct includes aOneshotAsyncSender<M> and in Kanal that type is !Sync. Since Ask<Q,A> is part of the respective MsgType that in turn is part of ActorHandle<MsgType> the whole ActorHandle was !Sync. Since hself:ActorHandle was passed into the async receive(&mut self, msg:MsgType, hself: &ActorHandle<MsgType>, ..) this required significant overhead (Pin<Box<Future..>> to make the Ask pattern work and also needed additional MsgType constraints in the user code. Both are bad.

Now Actor is a struct that owns both hself and hsys, i.e. there is no more need to pass them into the receive future and hence ActorHandle<MsgType> does not need to be Sync. Existence of a ActorReceive trait (which defines the receive() method) is guaranteed by separating actor creation (which instantiates the Actor<MsgType,State>) and actor spawning (which receives an A:ActorReceive<MsgType> argument). Since that happens at the user code level the compiler sees respective ActorReceive impls.

This has the drawback that user code types can be a little confusing - impl ActorReceive<MsgType> for Actor<MsgType,State> - and the only user code level type is State. This was alleviated by proc macros (define_actor_msg_type, impl_actor, spawn_actor) that mostly hide Actor and ActorReceive (Actor has blanket Deref and DerefMut impls). As a silver lining we now have a type we can attach system fields and methods to. This needs documentation though to explain the type magic we use to reduce boiler plate