refactored Ask into more general Query and added a QueryBuilder in case queries are repetitive (no need to create the channel per query). As it turns out there is nothing to be gained from kanal::oneshot so we drop it to further reduce the difference between channel implementations
added a benchmark_query. At first this was somewhat disappointing since it was ony 20% faster than Akka's ask pattern, but then it turned out this is only if the query happens from tokio::main. Moving it into another task (no difference if actor or non-actor) turned out to be between 30-300 times faster (depending on how many cycles we run, i.e. how much time we give hotspot to optimize).

added a RawAskBench to race-core-test for comparison